refactor(core): move components in PDP to own files
